+++ src/contract/ContractProvider.ts
@@ -6,9 +6,15 @@ import { Maybe } from "../utils/maybe";
 import { ContractState } from "./ContractState";
 import { Sender } from './Sender';
 
+export type ContractGetMethodResult = {
+    stack: TupleReader;
+    gasUsed?: Maybe<bigint>;
+    logs?: Maybe<string>;
+}
+
 export interface ContractProvider {
     getState(): Promise<ContractState>;
-    get(name: string, args: TupleItem[]): Promise<{ stack: TupleReader }>;
+    get(name: string, args: TupleItem[]): Promise<ContractGetMethodResult>;
     external(message: Cell): Promise<void>;
     internal(via: Sender, args: { value: bigint | string, bounce?: Maybe<boolean>, sendMode?: SendMode, body?: Maybe<Cell | string> }): Promise<void>;
 }