+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Added
+- `address` helper function for easier address parsing
+- `ContractGetMethodResult` and added optional `gasUsed` and `logs` to it
+
 ## [0.46.0] - 2023-01-23
 ## Added
 - export the `OpenedContract` type